Feature/payment

+++ src/main/java/com/shinhan/peoch/card/CardStatementDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Data
 public class CardStatementDTO {
     // 금액
-    private Long amount;
+    private Integer amount;
     // 결제일
     private LocalDateTime paymentDate;
     // 결제 방법
@@ -21,7 +21,7 @@ public class CardStatementDTO {
     // 할부 회차
     private Integer installmentRound;
     // 혜택 할인 금액
-    private Long benefitDiscountAmount;
+    private Integer benefitDiscountAmount;
 
 }
